refactor: Separate tests by concern
