+++ breeze/tests/attributes_test.ts
@@ -1,0 +1,41 @@
+import { assertEquals } from "https://deno.land/std@0.142.0/testing/asserts.ts";
+
+import breeze from "../index.ts";
+
+Deno.test("attributes", async () => {
+  assertEquals(
+    await breeze({
+      component: {
+        element: "a",
+        attributes: { href: "testing", empty: "" },
+        children: "testing",
+      },
+    }),
+    '<a href="testing" empty>testing</a>',
+  );
+});
+
+Deno.test("undefined attributes", async () => {
+  assertEquals(
+    await breeze({
+      component: {
+        element: "a",
+        attributes: { href: undefined },
+        children: "testing",
+      },
+    }),
+    "<a>testing</a>",
+  );
+});
+
+Deno.test("attributes without children", async () => {
+  assertEquals(
+    await breeze({
+      component: {
+        element: "a",
+        attributes: { href: "testing" },
+      },
+    }),
+    '<a href="testing"></a>',
+  );
+});
