+++ breeze/tests/nesting_test.ts
@@ -1,0 +1,61 @@
+import { assertEquals } from "https://deno.land/std@0.142.0/testing/asserts.ts";
+
+import breeze from "../index.ts";
+
+const onlyChildren = { children: "testing" };
+const emptySpan = { element: "span" };
+const span = { element: "span", children: "testing" };
+
+Deno.test("empty element", async () => {
+  assertEquals(await breeze({ component: emptySpan }), "<span></span>");
+});
+
+Deno.test("array of elements", async () => {
+  assertEquals(
+    await breeze({ component: [emptySpan, emptySpan] }),
+    "<span></span><span></span>",
+  );
+});
+
+Deno.test("simple element", async () => {
+  assertEquals(await breeze({ component: span }), "<span>testing</span>");
+});
+
+Deno.test("children element", async () => {
+  assertEquals(await breeze({ component: onlyChildren }), "testing");
+});
+
+Deno.test("nested element", async () => {
+  assertEquals(
+    await breeze({ component: { element: "div", children: [span] } }),
+    "<div><span>testing</span></div>",
+  );
+});
+
+Deno.test("nested siblings", async () => {
+  assertEquals(
+    await breeze({ component: { element: "div", children: [span, span] } }),
+    "<div><span>testing</span><span>testing</span></div>",
+  );
+});
+
+Deno.test("nested children siblings", async () => {
+  assertEquals(
+    await breeze({
+      component: { element: "div", children: [onlyChildren, onlyChildren] },
+    }),
+    "<div>testingtesting</div>",
+  );
+});
+
+Deno.test("multi-level nesting", async () => {
+  assertEquals(
+    await breeze({
+      component: {
+        element: "div",
+        children: [{ element: "div", children: [span] }],
+      },
+    }),
+    "<div><div><span>testing</span></div></div>",
+  );
+});
